+++ src/main/java/org/sitmun/plugin/core/domain/ApplicationParameter.java
@@ -5,6 +5,7 @@ import org.hibernate.annotations.OnDeleteAction;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigInteger;
 
 @Entity
 @Table(name = "STM_PARAMAPP")
@@ -13,8 +14,8 @@ public class ApplicationParameter {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "STM_GENERATOR")
   @SequenceGenerator(name = "STM_GENERATOR", sequenceName = "STM_SEQ")
-  @Column(name = "PAP_CODIGO")
-  private long id;
+  @Column(name = "PAP_CODIGO", precision = 11)
+  private BigInteger id;
 
   @Column(name = "PAP_NOMBRE", length = 30)
   private String name;
@@ -31,11 +32,11 @@ public class ApplicationParameter {
   @JoinColumn(name = "PAP_CODAPP", foreignKey = @ForeignKey(name = "STM_PAP_FK_APP"))
   private Application application;
 
-  public long getId() {
+  public BigInteger getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(BigInteger id) {
     this.id = id;
   }
