+++ src/main/java/org/sitmun/plugin/core/domain/Log.java
@@ -1,6 +1,7 @@
 package org.sitmun.plugin.core.domain;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 import java.util.Date;
 
 @Entity
@@ -10,8 +11,8 @@ public class Log {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "STM_GENERATOR")
   @SequenceGenerator(name = "STM_GENERATOR", sequenceName = "STM_SEQ")
-  @Column(name = "LOG_CODIGO")
-  private long id;
+  @Column(name = "LOG_CODIGO", precision = 11)
+  private BigInteger id;
 
   @Column(name = "LOG_TIPO", length = 50)
   private String nombre;
@@ -39,14 +40,14 @@ public class Log {
   /**
    * @return the id
    */
-  public long getId() {
+  public BigInteger getId() {
     return id;
   }
 
   /**
    * @param id the id to set
    */
-  public void setId(long id) {
+  public void setId(BigInteger id) {
     this.id = id;
   }
