+++ src/main/java/org/sitmun/plugin/core/domain/Task.java
@@ -7,6 +7,7 @@ import org.springframework.hateoas.Resource;
 import org.springframework.hateoas.ResourceSupport;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
@@ -18,8 +19,8 @@ public class Task implements Identifiable {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "STM_GENERATOR")
   @SequenceGenerator(name = "STM_GENERATOR", sequenceName = "STM_SEQ")
-  @Column(name = "TAR_CODIGO")
-  private long id;
+  @Column(name = "TAR_CODIGO", precision = 11)
+  private BigInteger id;
 
   @Column(name = "TAR_NOMBRE", length = 50)
   private String name;
@@ -66,11 +67,11 @@ public class Task implements Identifiable {
     this.ui = ui;
   }
 
-  public Long getId() {
+  public BigInteger getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(BigInteger id) {
     this.id = id;
   }
