+++ src/main/java/org/sitmun/plugin/core/domain/TaskAvailability.java
@@ -2,6 +2,7 @@ package org.sitmun.plugin.core.domain;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigInteger;
 import java.util.Date;
 
 @Entity
@@ -11,8 +12,8 @@ public class TaskAvailability {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "STM_GENERATOR")
   @SequenceGenerator(name = "STM_GENERATOR", sequenceName = "STM_SEQ")
-  @Column(name = "DTA_CODIGO")
-  private long id;
+  @Column(name = "DTA_CODIGO", precision = 11)
+  private BigInteger id;
 
   @Column(name = "DTA_F_ALTA")
   private Date createdDate;
@@ -27,11 +28,11 @@ public class TaskAvailability {
   @NotNull
   private Task task;
 
-  public long getId() {
+  public BigInteger getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(BigInteger id) {
     this.id = id;
   }
