+++ src/main/java/org/sitmun/plugin/core/domain/TaskGroup.java
@@ -1,6 +1,7 @@
 package org.sitmun.plugin.core.domain;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
 @Table(name = "STM_GRPTAR")
@@ -9,8 +10,8 @@ public class TaskGroup {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "STM_GENERATOR")
   @SequenceGenerator(name = "STM_GENERATOR", sequenceName = "STM_SEQ")
-  @Column(name = "GTA_CODIGO")
-  private long id;
+  @Column(name = "GTA_CODIGO", precision = 11)
+  private BigInteger id;
 
   @Column(name = "GTA_NOMBRE", length = 80)
   private String name;
@@ -18,14 +19,14 @@ public class TaskGroup {
   /**
    * @return the id
    */
-  public long getId() {
+  public BigInteger getId() {
     return id;
   }
 
   /**
    * @param id the id to set
    */
-  public void setId(long id) {
+  public void setId(BigInteger id) {
     this.id = id;
   }
