+++ src/main/java/org/sitmun/plugin/core/domain/TaskType.java
@@ -1,6 +1,7 @@
 package org.sitmun.plugin.core.domain;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
 @Table(name = "STM_TIPOTAREA")
@@ -9,17 +10,17 @@ public class TaskType {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "STM_GENERATOR")
   @SequenceGenerator(name = "STM_GENERATOR", sequenceName = "STM_SEQ")
-  @Column(name = "TTA_CODIGO")
-  private long id;
+  @Column(name = "TTA_CODIGO", precision = 11)
+  private BigInteger id;
 
   @Column(name = "TTA_NOMBRE", length = 30)
   private String name;
 
-  public long getId() {
+  public BigInteger getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(BigInteger id) {
     this.id = id;
   }
