+++ src/main/java/org/sitmun/plugin/core/domain/TaskUI.java
@@ -1,6 +1,7 @@
 package org.sitmun.plugin.core.domain;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
 @Table(name = "STM_TAREA_UI")
@@ -9,8 +10,8 @@ public class TaskUI {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "STM_GENERATOR")
   @SequenceGenerator(name = "STM_GENERATOR", sequenceName = "STM_SEQ")
-  @Column(name = "TUI_CODIGO")
-  private long id;
+  @Column(name = "TUI_CODIGO", precision = 11)
+  private BigInteger id;
 
   @Column(name = "TUI_NOMBRE", length = 30)
   private String name;
@@ -50,11 +51,11 @@ public class TaskUI {
   }
 
 
-  public long getId() {
+  public BigInteger getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(BigInteger id) {
     this.id = id;
   }
