+++ src/main/java/org/sitmun/plugin/core/domain/TreeNode.java
@@ -3,6 +3,7 @@ package org.sitmun.plugin.core.domain;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
 @Table(name = "STM_ARBOLNOD")
@@ -12,8 +13,8 @@ public class TreeNode {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "STM_GENERATOR")
   @SequenceGenerator(name = "STM_GENERATOR", sequenceName = "STM_SEQ")
-  @Column(name = "ANR_CODIGO")
-  private long id;
+  @Column(name = "ANR_CODIGO", precision = 11)
+  private BigInteger id;
 
   @Column(name = "ARN_NOMBRE", length = 80)
   private String name;
@@ -40,11 +41,11 @@ public class TreeNode {
   @ManyToOne
   private Cartography cartography;
 
-  public long getId() {
+  public BigInteger getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(BigInteger id) {
     this.id = id;
   }
