+++ src/main/java/org/sitmun/plugin/core/repository/LogRepository.java
@@ -4,8 +4,10 @@ import org.sitmun.plugin.core.domain.Log;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
+import java.math.BigInteger;
+
 @RepositoryRestResource(collectionResourceRel = "logs", path = "logs")
-public interface LogRepository extends PagingAndSortingRepository<Log, Long> {
+public interface LogRepository extends PagingAndSortingRepository<Log, BigInteger> {
 
 
 }
